PPTFormatReader fix по результатам тестирования (linux)
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@65254 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficePPTFile/PPTFormatLib/PPTFormatLib.h
@@ -25,7 +25,7 @@ public:
 
 private:
 
-	LONG m_Status;
+	int m_Status;
 
 	void* m_pReader;
